requirements.txt
Flask==2.3.3
yt-dlp==2025.9.10
openai-whisper==20230314
